+++ NEWS.md
@@ -77,6 +77,20 @@
 
 ## Dependency changes
 
+* none
+
+## Bugfixes
+
+* none
+
+## New features & enhancements
+
+* none
+
+# grainscape 0.5.0
+
+## Dependency changes
+
 * Require R 4.2 or higher (for native pipe and placeholder)
 * Add `DiagrammeR`, `dplyr`, `webshot2` and `withr` to Suggests for use with vignettes and tests